% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_srf.R
\name{compute_srf}
\alias{compute_srf}
\title{compute_srf}
\usage{
compute_srf(Waves, Values, SensorRSR, Group = NULL)
}
\arguments{
\item{Waves}{wavelengths of the original spectroradiometric values}

\item{Values}{sepctroradiometric values (radiance, irradiance, reflectance)}

\item{SensorRSR}{Sensor RSR as outputted by \link[SatRSR]{sensor_rsr}}

\item{ID}{}
}
\value{
A long format data frame, with columns \code{Sensor}, \code{SensorWavelength}, \code{SensorValues},
}
\description{
Transform spectral data to the response of a sensor.
Originally writted by Simon Bélanger, re-implemented by Raphaël mabit
}
