% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_srf.R
\name{sensor_rsr}
\alias{sensor_rsr}
\title{sensor_rsr}
\usage{
sensor_rsr(Sensor = "WISE")
}
\arguments{
\item{Sensor}{sensor name for which to fetch the RSR.
Available are: "WISE".}
}
\value{
a data frame of RSR with structure, Wavelength, \emph{Sensor Band Central Wavelength}{1,n}
}
\description{
Fetch the Relative Spectral Resonse of a sensor in the h5 database
}
\author{
Raphaël Mabit
}
